Move asserts to helpers
