+++ testing/tests/DevExpress.ui.widgets/drawer.scenarios.tests.js
@@ -214,88 +214,6 @@ configs.forEach(config => {
             return extend({ rtlEnabled: false, animationEnabled: false }, config, targetOptions);
         }
 
-        function checkWhenPanelContentRendered(assert, drawer, drawerElement, panelTemplateElement) {
-            const drawerRect = drawerElement.getBoundingClientRect();
-            if(config.minSize) {
-                if(drawer.option('openedStateMode') === 'overlap') {
-                    const panelRect = panelTemplateElement.parentElement.getBoundingClientRect();
-                    const expectedPanelRect = extend({}, drawerRect);
-                    if(drawer.option('minSize')) {
-                        switch(drawer.option('position')) {
-                            case 'left':
-                                expectedPanelRect.width = drawer.option('minSize');
-                                break;
-                            case 'right':
-                                expectedPanelRect.left += drawerRect.wight - drawer.option('minSize');
-                                expectedPanelRect.width = drawer.option('minSize');
-                                break;
-                            case 'top':
-                                expectedPanelRect.height = drawer.option('minSize');
-                                break;
-                            case 'bottom':
-                                expectedPanelRect.top = drawerRect.bottom - drawer.option('minSize');
-                                expectedPanelRect.height = drawer.option('minSize');
-                                break;
-                        }
-                    }
-                    assert.strictEqual(
-                        panelRect.left === expectedPanelRect.left
-                        && panelRect.top === expectedPanelRect.top
-                        && panelRect.width === expectedPanelRect.width
-                        && panelRect.height === expectedPanelRect.height,
-                        true,
-                        'panelRect/expectedRect: ' +
-                        `left:[${panelRect.left}/${expectedPanelRect.left}], top:[${panelRect.top}/${expectedPanelRect.top}], ` +
-                        `width:[${panelRect.width}/${expectedPanelRect.width}], height:[${panelRect.height}/${expectedPanelRect.height}]`);
-                }
-            } else {
-                if(drawer.option('openedStateMode') !== 'overlap') {
-                    const panelRect = panelTemplateElement.parentElement.getBoundingClientRect();
-                    assert.strictEqual(
-                        panelRect.right < drawerRect.left
-                        || panelRect.left > drawerRect.right
-                        || panelRect.bottom < drawerRect.top
-                        || panelRect.top > drawerRect.bottom,
-                        true,
-                        'panel should be out of drawerRect, ' +
-                        `left:[${panelRect.left}/${drawerRect.left}], top:[${panelRect.top}/${drawerRect.top}], ` +
-                        `right:[${panelRect.right}/${drawerRect.right}], bottom:[${panelRect.bottom}/${drawerRect.bottom}], [panel/drawer]`);
-                }
-            }
-
-            if(!drawer.option('minSize') || (drawer.option('openedStateMode') === 'overlap')) {
-                const expectedViewRect = extend({}, drawerRect);
-                if(drawer.option('minSize')) {
-                    switch(drawer.option('position')) {
-                        case 'left':
-                            expectedViewRect.left += drawer.option('minSize');
-                            expectedViewRect.width -= drawer.option('minSize');
-                            break;
-                        case 'right':
-                            expectedViewRect.width -= drawer.option('minSize');
-                            break;
-                        case 'top':
-                            expectedViewRect.top += drawer.option('minSize');
-                            expectedViewRect.height -= drawer.option('minSize');
-                            break;
-                        case 'bottom':
-                            expectedViewRect.height -= drawer.option('minSize');
-                            break;
-                    }
-                }
-                const viewRect = document.getElementById('view').getBoundingClientRect();
-                assert.strictEqual(
-                    viewRect.left === expectedViewRect.left
-                    && viewRect.top === expectedViewRect.top
-                    && viewRect.width === expectedViewRect.width
-                    && viewRect.height === expectedViewRect.height,
-                    true,
-                    'viewRect/expectedRect, ' +
-                    `left:[${viewRect.left}/${expectedViewRect.left}], top:[${viewRect.top}/${expectedViewRect.top}], ` +
-                    `width:[${viewRect.width}/${expectedViewRect.width}], height:[${viewRect.height}/${expectedViewRect.height}]`);
-            }
-        }
-
         testOrSkip('opened: false', () => configIs('push', 'top') || configIs('overlap', ['right', 'top'], ['expand', 'slide']) && config.minSize, function(assert) {
             const drawerElement = document.getElementById(drawerTesters.drawerElementId);
 
@@ -303,7 +221,7 @@ configs.forEach(config => {
                 opened: false,
                 template: drawerTesters[config.position].template,
                 __debugWhenPanelContentRendered: (e) => {
-                    checkWhenPanelContentRendered(assert, e.drawer, drawerElement, document.getElementById('template'));
+                    drawerTesters[config.position].checkWhenPanelContentRendered(assert, e.drawer, drawerElement, document.getElementById('template'), config.minSize);
                 }
             }));
